requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Queue
redis==5.0.1
rq==1.15.1

# Browser Automation
playwright==1.41.0

# Data Processing
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2

# Data Generation (for missing fields)
faker==22.0.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP & API
httpx==0.26.0
aiohttp==3.9.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Logging & Monitoring
structlog==24.1.0
psutil==5.9.7

# Utilities
tenacity==8.2.3
phonenumbers==8.13.27

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
mypy==1.8.0

# ASGI Server
gunicorn==21.2.0
